add users to db, secure db changing

+++ WebApplicationDemo/WebApp/Controllers/ExchangeKeysController.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Domain;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
 using WebApp.Data;
 
 namespace WebApp.Controllers
@@ -18,11 +20,18 @@ namespace WebApp.Controllers
         {
             _context = context;
         }
+        
+        string GetUserId()
+        {
+            var claim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+            return claim.Value ?? "";
+        }
 
         // GET: ExchangeKeys
         public async Task<IActionResult> Index()
         {
-            return View(await _context.ExchangeKeys.ToListAsync());
+            var userId = GetUserId();
+            return View(await _context.ExchangeKeys.Where(c =>c.UserId == userId).ToListAsync());
         }
 
         // GET: ExchangeKeys/Details/5
@@ -32,6 +41,9 @@ namespace WebApp.Controllers
             {
                 return NotFound();
             }
+            
+            if (!IdExists(id))
+                return RedirectToAction("Index");
 
             var exchangeKey = await _context.ExchangeKeys
                 .FirstOrDefaultAsync(m => m.Id == id);
@@ -48,7 +60,7 @@ namespace WebApp.Controllers
         {
             return View();
         }
-
+        
         // POST: ExchangeKeys/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
@@ -56,9 +68,15 @@ namespace WebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(ExchangeKey exchangeKey)
         {
+            if (!Crypto.KeyExchange.PrimesValidation(exchangeKey.G))
+                ModelState.AddModelError("G","G must be a prime number");
+            if (!Crypto.KeyExchange.PrimesValidation(exchangeKey.P))
+                ModelState.AddModelError("P","P must be a prime number");
+
             if (ModelState.IsValid)
             {
                 Crypto.KeyExchange.KeyExchangeImplementation(exchangeKey);
+                exchangeKey.UserId = GetUserId();
                 _context.Add(exchangeKey);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -73,6 +91,9 @@ namespace WebApp.Controllers
             {
                 return NotFound();
             }
+            
+            if (!IdExists(id))
+                return RedirectToAction("Index");
 
             var exchangeKey = await _context.ExchangeKeys.FindAsync(id);
             if (exchangeKey == null)
@@ -87,17 +108,25 @@ namespace WebApp.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,P,G,ASecret,BSecret,CommonSecret")] ExchangeKey exchangeKey)
+        public async Task<IActionResult> Edit(int id, ExchangeKey exchangeKey)
         {
             if (id != exchangeKey.Id)
             {
                 return NotFound();
             }
+            if (!IdExists(id))
+                return RedirectToAction("Index");
 
+            if (!Crypto.KeyExchange.PrimesValidation(exchangeKey.G))
+                ModelState.AddModelError("G","G must be a prime number");
+            if (!Crypto.KeyExchange.PrimesValidation(exchangeKey.P))
+                ModelState.AddModelError("P","P must be a prime number");
+            
             if (ModelState.IsValid)
             {
                 try
                 {
+                    exchangeKey.UserId = GetUserId();
                     _context.Update(exchangeKey);
                     await _context.SaveChangesAsync();
                 }
@@ -124,6 +153,8 @@ namespace WebApp.Controllers
             {
                 return NotFound();
             }
+            if (!IdExists(id))
+                return RedirectToAction("Index");
 
             var exchangeKey = await _context.ExchangeKeys
                 .FirstOrDefaultAsync(m => m.Id == id);
@@ -140,6 +171,8 @@ namespace WebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!IdExists(id))
+                return RedirectToAction("Index");
             var exchangeKey = await _context.ExchangeKeys.FindAsync(id);
             _context.ExchangeKeys.Remove(exchangeKey);
             await _context.SaveChangesAsync();
@@ -149,6 +182,10 @@ namespace WebApp.Controllers
         private bool ExchangeKeyExists(int id)
         {
             return _context.ExchangeKeys.Any(e => e.Id == id);
+        }
+        private bool IdExists(int ?id)
+        {
+            return _context.RsaKeys.Any(c => c.Id == id && c.UserId == GetUserId());
         }
     }
 }
