+++ WebApplicationDemo/WebApp/Controllers/RsaKeysController.cs
@@ -1,15 +1,23 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.Eventing.Reader;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Crypto;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Domain;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc.Diagnostics;
+using Microsoft.AspNetCore.Routing.Matching;
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using WebApp.Data;
 
 namespace WebApp.Controllers
 {
+    [Authorize]
     public class RsaKeysController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -19,15 +27,31 @@ namespace WebApp.Controllers
             _context = context;
         }
 
+        string GetUserId()
+        {
+           var claim = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+           return claim.Value ?? "";
+        }
+        
         // GET: RsaKeys
         public async Task<IActionResult> Index()
         {
-            return View(await _context.RsaKeys.ToListAsync());
+            var userId = GetUserId();
+            return View(await _context.RsaKeys.Where(c =>c.UserId == userId).ToListAsync());
         }
 
         // GET: RsaKeys/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            try
+            {
+                _context.RsaKeys.First(c => c.Id == id && c.UserId == GetUserId());
+            }
+            catch (InvalidOperationException)
+            {
+                return RedirectToAction("Index");
+            }
+            
             if (id == null)
             {
                 return NotFound();
@@ -59,6 +83,7 @@ namespace WebApp.Controllers
             if (ModelState.IsValid)
             {
                 Crypto.Rsa.RsaImplemantation(rsaKey);
+                rsaKey.UserId = GetUserId();
                 _context.Add(rsaKey);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -82,6 +107,7 @@ namespace WebApp.Controllers
             if (ModelState.IsValid)
             {
                 Crypto.Rsa.RsaDecrypt(rsaKey);
+                rsaKey.UserId = GetUserId();
                 _context.Add(rsaKey);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -92,12 +118,17 @@ namespace WebApp.Controllers
         // GET: RsaKeys/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+
             if (id == null)
             {
                 return NotFound();
             }
+            
+            if (!IdExists(id))
+                return RedirectToAction("Index");
 
             var rsaKey = await _context.RsaKeys.FindAsync(id);
+
             if (rsaKey == null)
             {
                 return NotFound();
@@ -116,12 +147,16 @@ namespace WebApp.Controllers
             {
                 return NotFound();
             }
-
+            
+            if (!IdExists(id))
+                return RedirectToAction("Index");
+            
             if (ModelState.IsValid)
             {
                 try
                 {
                     Crypto.Rsa.RsaImplemantation(rsaKey);
+                    rsaKey.UserId = GetUserId();
                     _context.Update(rsaKey);
                     await _context.SaveChangesAsync();
                 }
@@ -141,14 +176,21 @@ namespace WebApp.Controllers
             return View(rsaKey);
         }
 
+        
+
         // GET: RsaKeys/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            
+            
             if (id == null)
             {
                 return NotFound();
             }
 
+            if (!IdExists(id))
+                return RedirectToAction("Index");
+            
             var rsaKey = await _context.RsaKeys
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (rsaKey == null)
@@ -164,6 +206,8 @@ namespace WebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!IdExists(id))
+                return RedirectToAction("Index");
             var rsaKey = await _context.RsaKeys.FindAsync(id);
             _context.RsaKeys.Remove(rsaKey);
             await _context.SaveChangesAsync();
@@ -173,6 +217,11 @@ namespace WebApp.Controllers
         private bool RsaKeyExists(int id)
         {
             return _context.RsaKeys.Any(e => e.Id == id);
+        }
+
+        private bool IdExists(int ?id)
+        {
+            return _context.RsaKeys.Any(c => c.Id == id && c.UserId == GetUserId());
         }
     }
 }
